mypo-64 remove relative path
