+++ mypo/optimizer/__init__.py
@@ -1,3 +1,5 @@
 # flake8: noqa
 
-from .optimizer import MinimumVarianceOptimizer, SharpRatioOptimizer
+from mypo.optimizer.minimum_variance_optimizer import MinimumVarianceOptimizer
+from mypo.optimizer.optimizer import Optimizer
+from mypo.optimizer.sharp_ratio_optimizer import SharpRatioOptimizer
