+++ mypo/rebalancer/__init__.py
@@ -1,3 +1,6 @@
 # flake8: noqa
 
-from .rebalancer import MonthlyRebalancer, PlainRebalancer, ThresholdRebalancer
+from mypo.rebalancer.monthly_rebalancer import MonthlyRebalancer
+from mypo.rebalancer.plain_rebalancer import PlainRebalancer
+from mypo.rebalancer.rebalancer import Rebalancer
+from mypo.rebalancer.threshold_rebalancer import ThresholdRebalancer
