+++ mypo/rebalancer/plain_rebalancer.py
@@ -4,12 +4,26 @@ import datetime
 import numpy as np
 import numpy.typing as npt
 
+from mypo.common import safe_cast
+from mypo.rebalancer.rebalancer import Rebalancer
 
-class Rebalancer(object):
-    """Interface class of Rebalance stragegy class."""
 
-    def __init__(self) -> None:
-        pass
+class PlainRebalancer(Rebalancer):
+    """Simple weighted rebalance strategy."""
+
+    _weights: np.ndarray
+
+    def __init__(self, weights: npt.ArrayLike) -> None:
+        """
+        Construct object.
+
+        Parameters
+        ----------
+        weights
+            Weight for applying rebalance.
+        """
+        super().__init__()
+        self._weights = safe_cast(weights)
 
     def apply(
         self, index: datetime.datetime, assets: npt.ArrayLike, cash: np.float64
@@ -32,4 +46,5 @@ class Rebalancer(object):
         -------
         Deal
         """
-        pass
+        diff: np.ndarray = self._weights * np.sum(assets) - safe_cast(assets)
+        return diff
